Beginning the rgteaching refactor. Introducing instant workshop, where users can instantly create and fill out a workshop stats without creating an event first
