+++ .gitignore
@@ -13,3 +13,6 @@ myrobogals.db
 *.swp
 *.pyc
 *.o
+.pylintrc
+.idea/
+.vscode/
