+++ rgtemplates/includes/header.html
@@ -207,7 +207,19 @@
         </a>
         
         
+</li>
+
+                    <li class="item71" >
+            <a class="orphan item bullet" href="/teaching/instantworkshop/"  >
+            <span>
+                        {% trans "Create instant workshop" %}               
+            </span>
+
+        </a>
+        
+        
 </li>   
+
                     <li class="item71" >
             <a class="orphan item bullet" href="/teaching/schools/"  >
             <span>
