+++ rgtemplates/instant_workshop.html
@@ -1,0 +1,13 @@
+{% extends "base.html" %}
+{% load i18n %}
+{% block title %}{% trans "Create Instant Workshop" %}{% endblock %}
+
+{% block pagetitle %}{% trans "Create Instant Workshop" %}{% endblock %}
+
+{% block content %}
+<form action="/teaching/instantworkshop/" method="POST">{% csrf_token %}
+    {{ form1.as_p }}
+    {{ form2.as_p}}
+    <input type="submit" value="Next" />
+</form>
+{% endblock %}
